pharma_back/requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
flake8==6.1.0
idna==3.11
inflection==0.5.1
isort==5.13.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy_extensions==1.1.0
mysql-connector==2.2.9
pymysql==1.1.0
numpy==2.3.3
openpyxl==3.1.2
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xlrd==2.0.1
gunicorn==21.2.0
whitenoise==6.6.0
